refactor: 도메인, ui, validator 패키지 분리
