+++ src/main/java/lotto/Application.java
@@ -1,6 +1,6 @@
 package lotto;
 
-import java.util.List;
+import lotto.domain.LottoGame;
 
 public class Application {
     public static void main(String[] args) {
