+++ src/main/java/lotto/domain/LottoGame.java
@@ -1,6 +1,8 @@
-package lotto;
+package lotto.domain;
 
 import camp.nextstep.edu.missionutils.Randoms;
+import lotto.ui.Input;
+import lotto.ui.Output;
 
 import java.util.ArrayList;
 import java.util.List;
