+++ src/main/java/lotto/ui/Output.java
@@ -1,4 +1,6 @@
-package lotto;
+package lotto.ui;
+
+import lotto.domain.Lotto;
 
 import java.util.List;
