+++ src/main/java/lotto/validator/Validator.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.validator;
 
 import java.util.List;
